+++ packages/waggle/src/common/types.ts
@@ -59,6 +59,7 @@ export class ConnectionsManagerOptions {
   torControlPort: number
   torPassword?: string
   torAuthCookie?: string
+  useLocalTorFiles?: boolean = false // Use tor binaries and their libs from waggle
 }
 
 export interface IConstructor {