+++ packages/waggle/src/libp2p/connectionsManager.ts
@@ -19,6 +19,7 @@ import initListeners from '../socket/listeners'
 import IOProxy from '../socket/IOProxy'
 import { Connection } from 'libp2p-gossipsub/src/interfaces'
 import { HttpsProxyAgent } from 'https-proxy-agent'
+import path from 'path'
 
 const log = Object.assign(debug('waggle:conn'), {
   error: debug('waggle:conn:err')
@@ -80,6 +81,7 @@ export class ConnectionsManager {
 
   public initListeners = () => {
     initListeners(this.io, new IOProxy(this))
+    log('Initialized socket listeners')
   }
 
   public createNetwork = async () => {
@@ -95,9 +97,17 @@ export class ConnectionsManager {
 
   public init = async () => {
     this.initListeners()
+    await this.spawnTor()
+  }
+
+  public spawnTor = async () => {
+    let basePath = ''
+    if (this.options.useLocalTorFiles) {
+      basePath = path.join(__dirname, '../..')
+    }
 
     this.tor = new Tor({
-      torPath: torBinForPlatform(),
+      torPath: torBinForPlatform(basePath),
       appDataPath: this.zbayDir,
       controlPort: this.options.torControlPort,
       socksPort: this.agentPort,
@@ -106,7 +116,7 @@ export class ConnectionsManager {
 
       options: {
         env: {
-          LD_LIBRARY_PATH: torDirForPlatform(),
+          LD_LIBRARY_PATH: torDirForPlatform(basePath),
           HOME: os.homedir()
         },
         detached: true
@@ -115,8 +125,10 @@ export class ConnectionsManager {
 
     if (this.options.spawnTor) {
       await this.tor.init()
+      log('Spawned Tor')
     } else {
       this.tor.initTorControl()
+      log('Initialized tor control')
     }
   }