+++ packages/waggle/CHANGELOG.md
@@ -2,11 +2,16 @@
 
 ## [Unreleased]
 
+### Fixed
+
+* Stucking Tor - kill and relaunch the Tor after given timeout
+
 ### Added
 
 * DestroyHiddenService method in torManager
 * CreateNetwork in communitiesManager
 * CreateNetwork api in IOPRoxy
+* ConnectionsManager option - allow using tor files (binaries and libs) from waggle
 
 ## [4.0.1]